requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.1.1
distlib==0.3.4
Django==3.2
django-cors-headers==3.13.0
ffmpeg-python==0.2.0
filelock==3.6.0
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
idna==3.4
mercurial==5.6.1
multidict==6.0.4
openai==0.27.8
packaging==23.1
Pillow==9.0.0
pipenv==2022.4.21
platformdirs==2.5.2
psutil==5.9.5
psycopg2-binary==2.9.3
pydantic==1.10.11
pytz==2020.1
replicate==0.8.4
requests==2.28.1
roman==3.3
six==1.16.0
sqlparse==0.4.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==1.25.11
virtualenv==20.14.1
virtualenv-clone==0.5.7
Wand==0.6.10
yarl==1.9.2
